refactor: Migrate community plugins to a unified plugin registry
